+++ config/storybook/decorators/ThemeDecorator.tsx
@@ -1,4 +1,4 @@
 import { Story } from '@storybook/react';
-import { Theme } from '../../../src/01-app/providers/ThemeProvider';
+import { Theme } from '../../../src/07-shared/lib/components/ThemeProvider';
 
 export const ThemeDecorator = (theme: Theme) => (StoryComponent: Story) => <div className={`app ${theme}`}><StoryComponent /></div>;
