+++ src/05-features/ThemeSwitcher/ui/ThemeSwitcher.tsx
@@ -1,4 +1,4 @@
-import { Theme, useTheme } from '01-app/providers/ThemeProvider';
+import { Theme, useTheme } from '07-shared/lib/components/ThemeProvider';
 import SunIcon from '07-shared/assets/icons/sun.svg';
 import MoonIcon from '07-shared/assets/icons/moon.svg';
 import { classNames } from '07-shared/lib/classNames/classNames';
