+++ src/07-shared/lib/components/ErrorBoundary/index.ts
