+++ src/07-shared/lib/components/ErrorBoundary/ui/ErrorBoundary.tsx
